flask/requirements.txt
azure-monitor-opentelemetry==1.0.0b14
backoff==2.2.1
blinker==1.6.2
certifi==2023.5.7
charset-normalizer==3.2.0
click==8.1.4
Deprecated==1.2.14
Flask==2.3.2
googleapis-common-protos==1.59.1
grpcio==1.56.0
idna==3.4
importlib-metadata==6.0.1
install==1.3.5
itsdangerous==2.1.2
Jinja2==3.1.2
MarkupSafe==2.1.3
opentelemetry-api==1.18.0
opentelemetry-exporter-otlp==1.18.0
opentelemetry-exporter-otlp-proto-common==1.18.0
opentelemetry-exporter-otlp-proto-grpc==1.18.0
opentelemetry-exporter-otlp-proto-http==1.18.0
opentelemetry-instrumentation==0.39b0
opentelemetry-instrumentation-flask==0.39b0
opentelemetry-instrumentation-requests==0.39b0
opentelemetry-instrumentation-wsgi==0.39b0
opentelemetry-instrumentation-urllib3==0.39b0
opentelemetry-proto==1.18.0
opentelemetry-sdk==1.18.0
opentelemetry-semantic-conventions==0.39b0
opentelemetry-util-http==0.39b0
protobuf==4.23.4
requests==2.31.0
typing_extensions==4.7.1
urllib3==2.0.3
Werkzeug==2.3.6
wrapt==1.15.0
zipp==3.16.0
